合并报错提交(修改未成功)

+++ src/main/java/com/zy/pojo/Renttable.java
@@ -49,7 +49,7 @@ public class Renttable {
 	/**
 	 * 是否归还
 	 */
-	private Double rentflag;
+	private Integer rentflag;
 
 	/**
 	 * 身份证
